defined props for footer
